+++ WalletWasabi.Fluent/ViewModels/Wallets/CoinJoinStateViewModel.cs
@@ -9,6 +9,7 @@ using ReactiveUI;
 using WalletWasabi.Fluent.Helpers;
 using WalletWasabi.Fluent.State;
 using WalletWasabi.WabiSabi.Client;
+using WalletWasabi.WabiSabi.Client.Events;
 using WalletWasabi.Wallets;
 
 namespace WalletWasabi.Fluent.ViewModels.Wallets;
