+++ WalletWasabi/WabiSabi/Client/Events/CompletedEventArgs.cs
@@ -1,0 +1,14 @@
+using WalletWasabi.Wallets;
+
+namespace WalletWasabi.WabiSabi.Client.Events;
+
+public class CompletedEventArgs : StatusChangedEventArgs
+{
+	public CompletedEventArgs(Wallet wallet, CompletionStatus completionStatus)
+		: base(wallet)
+	{
+		CompletionStatus = completionStatus;
+	}
+
+	public CompletionStatus CompletionStatus { get; }
+}
