+++ WalletWasabi/WabiSabi/Client/Events/LoadedEventArgs.cs
@@ -1,0 +1,11 @@
+using WalletWasabi.Wallets;
+
+namespace WalletWasabi.WabiSabi.Client.Events;
+
+public class LoadedEventArgs : StatusChangedEventArgs
+{
+	public LoadedEventArgs(Wallet wallet)
+		: base(wallet)
+	{
+	}
+}
