+++ WalletWasabi/WabiSabi/Client/Events/StatusChangedEventArgs.cs
@@ -1,0 +1,34 @@
+using WalletWasabi.Wallets;
+
+namespace WalletWasabi.WabiSabi.Client.Events;
+
+public enum StopReason
+{
+	WalletUnloaded
+}
+
+public enum CompletionStatus
+{
+	Success,
+	Canceled,
+	Failed,
+	Unknown
+}
+
+public enum CoinjoinError
+{
+	NoCoinsToMix,
+	AutoConjoinDisabled,
+	UserInSendWorkflow,
+	NotEnoughUnprivateBalance
+}
+
+public class StatusChangedEventArgs : EventArgs
+{
+	public StatusChangedEventArgs(Wallet wallet)
+	{
+		Wallet = wallet;
+	}
+
+	public Wallet Wallet { get; }
+}
